feat: ErrorBoundary 에서 공용적으로 이용될 CommonErrorBoundary로 변경
